% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimates.R
\name{compute_bias_risk_ratio}
\alias{compute_bias_risk_ratio}
\title{Compute the relative bias in the estimated risk ratio as a percentage of the true value}
\usage{
compute_bias_risk_ratio(estimated_risk_ratio, true_risk_ratio)
}
\arguments{
\item{estimated_risk_ratio}{The estimated risk ratio for the
association of exposure with the outcome returned by any of the
\verb{run_*} analysis functions.}

\item{true_risk_ratio}{The true risk ratio for the association of
exposure with the outcome.}
}
\value{
The percent relative bias of the estimated risk ratio.
}
\description{
Compute the relative bias in the estimated risk ratio as a percentage of the true value
}
\examples{
probability_exposed <- 0.5
probability_stratum_1_given_exposed <- 0.6
probability_stratum_1_given_unexposed <- 0.4
risk_given_unexposed_stratum_2 <- 0.2
sim_data <- simulate_data(
  n = 100,
  probability_exposed,
  probability_stratum_1_given_exposed,
  probability_stratum_1_given_unexposed,
  risk_given_unexposed_stratum_2 = 0.2,
  relative_risk_exposed = 2,
  relative_risk_stratum_1 = 1.5
)
poisson_glm_fit <- run_poisson_glm(
  outcome ~ exposed + stratum, data = sim_data
)
compute_bias_risk_ratio(poisson_glm_fit$est, 2)
}
