% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{compute_risk}
\alias{compute_risk}
\title{Compute the risk of the outcome given an observation's exposure and stratum along with the baseline risk and the relative risks of exposure and belonging to stratum 1.}
\usage{
compute_risk(
  exposure,
  stratum,
  risk_given_unexposed_stratum_2,
  relative_risk_exposed,
  relative_risk_stratum_1
)
}
\arguments{
\item{exposure}{A vector of exposure/treatment indicators.  Assumes
that 1 represents exposed/treated and 0 represents
unexposed/control.}

\item{stratum}{A vector of stratum membership indicators.  Assumes
that strata are coded 1 and 2.}

\item{risk_given_unexposed_stratum_2}{The risk of the outcome for
observations that are unexposed and in the second stratum.}

\item{relative_risk_exposed}{The relative risk of the outcome
associated with exposure.}

\item{relative_risk_stratum_1}{The relative risk of the outcome
associated with the first stratum.}
}
\value{
The risk, i.e., the probability of experiencing the
outcome.
}
\description{
Compute the risk of the outcome given an observation's exposure and stratum along with the baseline risk and the relative risks of exposure and belonging to stratum 1.
}
\examples{
risk_given_unexposed_stratum_2 <- 0.2
relative_risk_exposed <- 2
relative_risk_stratum_1 <- 1.5
data <- expand.grid(
  exposed = 1:0,
  stratum = 1:2
)
compute_risk(
  data$exposed,
  data$stratum,
  risk_given_unexposed_stratum_2,
  relative_risk_exposed,
  relative_risk_stratum_1
)
}
