% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_results.R
\name{get_result_by_approach}
\alias{get_result_by_approach}
\title{Get a component of the results from a single simulation run for a specified approach from a list of simulation results}
\usage{
get_result_by_approach(simulation_results, result, approach)
}
\arguments{
\item{simulation_results}{A list of simulation run results produced
by \code{run_simulation}}

\item{result}{A string matching the name of one of the entries of
the list returned by any of the \verb{run_*} functions.}

\item{approach}{A string matching the name of one of the analysis
approaches.}
}
\value{
A vector of the results.
}
\description{
Get a component of the results from a single simulation run for a specified approach from a list of simulation results
}
\examples{
probability_exposed <- 0.5
probability_stratum_1_given_exposed <- 0.6
probability_stratum_1_given_unexposed <- 0.4
risk_given_unexposed_stratum_2 <- 0.2
results <- lapply(
  1:10,
  function () {
    run_simulation(
      n = 100,
      probability_exposed,
      probability_stratum_1_given_exposed,
      probability_stratum_1_given_unexposed,
      risk_given_unexposed_stratum_2 = 0.2,
      relative_risk_exposed = 2,
      relative_risk_stratum_1 = 1.5
    )
  }
)
get_result_by_approach(results, "est", "modified_poisson_glm")
}
