% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimates.R
\name{is_true_value_covered}
\alias{is_true_value_covered}
\title{Is the true risk ratio contained in the given interval?}
\usage{
is_true_value_covered(confidence_interval, true_risk_ratio)
}
\arguments{
\item{confidence_interval}{The confidence interval of the risk
ratio for the association of exposure with the outcome returned
by any of the \verb{run_*} analysis functions.}

\item{true_risk_ratio}{The true risk ratio for the association of
exposure with the outcome.}
}
\value{
A Boolean indicating whether the estimate is contained in
the interval.
}
\description{
Is the true risk ratio contained in the given interval?
}
\examples{
probability_exposed <- 0.5
probability_stratum_1_given_exposed <- 0.6
probability_stratum_1_given_unexposed <- 0.4
risk_given_unexposed_stratum_2 <- 0.2
sim_data <- simulate_data(
  n = 100,
  probability_exposed,
  probability_stratum_1_given_exposed,
  probability_stratum_1_given_unexposed,
  risk_given_unexposed_stratum_2 = 0.2,
  relative_risk_exposed = 2,
  relative_risk_stratum_1 = 1.5
)
poisson_glm_fit <- run_poisson_glm(
  outcome ~ exposed + stratum, data = sim_data
)
is_true_value_covered(poisson_glm_fit$ci, 2)
}
