% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplicateModifiedPoisson-package.R
\docType{package}
\name{ReplicateModifiedPoisson-package}
\alias{ReplicateModifiedPoisson}
\alias{ReplicateModifiedPoisson-package}
\title{ReplicateModifiedPoisson: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jsahrmann/ReplicateModifiedPoisson}
  \item Report bugs at \url{https://github.com/jsahrmann/ReplicateModifiedPoisson/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
