% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_results.R
\name{report_coverage_risk_ratio}
\alias{report_coverage_risk_ratio}
\title{Compute the empirical coverage percentage of the confidence
intervals for the estimated risk ratio across a series of
simulation runs}
\usage{
report_coverage_risk_ratio(simulation_results, true_risk_ratio)
}
\arguments{
\item{simulation_results}{A list of simulation run results produced
by \code{run_simulation}}

\item{true_risk_ratio}{The true risk ratio for the association of
exposure with the outcome used in the simulations.}
}
\value{
The empirical coverage percentage of the confidence
intervals for the estimated risk across the simulation runs.
}
\description{
Compute the empirical coverage percentage of the confidence
intervals for the estimated risk ratio across a series of
simulation runs
}
\examples{
probability_exposed <- 0.5
probability_stratum_1_given_exposed <- 0.6
probability_stratum_1_given_unexposed <- 0.4
risk_given_unexposed_stratum_2 <- 0.2
results <- lapply(
  1:10,
  function () {
    run_simulation(
      n = 100,
      probability_exposed,
      probability_stratum_1_given_exposed,
      probability_stratum_1_given_unexposed,
      risk_given_unexposed_stratum_2 = 0.2,
      relative_risk_exposed = 2,
      relative_risk_stratum_1 = 1.5
    )
  }
)
report_coverage_risk_ratio(results, 2)
}
