% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_results.R
\name{report_relative_bias_risk_ratio}
\alias{report_relative_bias_risk_ratio}
\title{Compute the relative bias in the estimated risk ratio as a
percentage of the true value across a series of simulation runs}
\usage{
report_relative_bias_risk_ratio(simulation_results, true_risk_ratio)
}
\arguments{
\item{simulation_results}{A list of simulation run results produced
by \code{run_simulation}}

\item{true_risk_ratio}{The true risk ratio for the association of
exposure with the outcome used in the simulations.}
}
\value{
The percent relative bias of the estimated risk ratio
across the simulation runs.
}
\description{
Compute the relative bias in the estimated risk ratio as a
percentage of the true value across a series of simulation runs
}
\examples{
probability_exposed <- 0.5
probability_stratum_1_given_exposed <- 0.6
probability_stratum_1_given_unexposed <- 0.4
risk_given_unexposed_stratum_2 <- 0.2
results <- lapply(
  1:10,
  function () {
    run_simulation(
      n = 100,
      probability_exposed,
      probability_stratum_1_given_exposed,
      probability_stratum_1_given_unexposed,
      risk_given_unexposed_stratum_2 = 0.2,
      relative_risk_exposed = 2,
      relative_risk_stratum_1 = 1.5
    )
  }
)
report_relative_bias_risk_ratio(results, 2)
}
