% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_analyses.R
\name{run_mantel_haenszel}
\alias{run_mantel_haenszel}
\title{Run the Mantel-Haenszel procedure estimating the risk ratio associated with a (potentially stratified) exposure/treatment}
\usage{
run_mantel_haenszel(exposure, outcome, stratum = NULL, conf_level = 0.95)
}
\arguments{
\item{exposure}{A vector of exposure/treatment indicators.  Assumes
that 1 represents exposed/treated and 0 represents
unexposed/control.}

\item{outcome}{A vector of outcome indicators.  Assumes that 1
represents the occurrence of the event of interest and 0
represents no event.}

\item{stratum}{A vector indicating stratum membership.  Default
\code{NULL} means the data are not stratified.}

\item{conf_level}{The confidence level of the interval.}
}
\value{
A list with elements \code{est} containing the estimated risk
ratio and \code{ci} containing the limits of the 100 * (1 -
conf_level)\% confidence interval.
}
\description{
Run the Mantel-Haenszel procedure estimating the risk ratio associated with a (potentially stratified) exposure/treatment
}
\examples{
probability_exposed <- 0.5
probability_stratum_1_given_exposed <- 0.6
probability_stratum_1_given_unexposed <- 0.4
risk_given_unexposed_stratum_2 <- 0.2
sim_data <- simulate_data(
  n = 100,
  probability_exposed,
  probability_stratum_1_given_exposed,
  probability_stratum_1_given_unexposed,
  risk_given_unexposed_stratum_2 = 0.2,
  relative_risk_exposed = 2,
  relative_risk_stratum_1 = 1.5
)
run_mantel_haenszel(
  exposure = sim_data$exposed,
  outcome = sim_data$outcome
)
run_mantel_haenszel(
  exposure = sim_data$exposed,
  outcome = sim_data$outcome,
  stratum = sim_data$stratum
)
run_mantel_haenszel(
  exposure = sim_data$exposed,
  outcome = sim_data$outcome,
  stratum = sim_data$stratum,
  conf_level = 0.99
)
}
