% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate a data set according to Zou (2004, p 703)}
\usage{
simulate_data(
  n,
  probability_exposed,
  probability_stratum_1_given_exposed,
  probability_stratum_1_given_unexposed,
  risk_given_unexposed_stratum_2,
  relative_risk_exposed,
  relative_risk_stratum_1
)
}
\arguments{
\item{n}{The number of observations in the data set.}

\item{probability_exposed}{The probability that an observation is
in the 'exposed' group}

\item{probability_stratum_1_given_exposed}{The conditional
probability that an observation is in the first stratum given
that the observation is in the exposed group.}

\item{probability_stratum_1_given_unexposed}{The conditional
probability that an observation is in the first stratum given
that the observation is in the unexposed group.}

\item{risk_given_unexposed_stratum_2}{The risk of the outcome for
observations that are unexposed and in the second stratum.}

\item{relative_risk_exposed}{The relative risk of the outcome
associated with exposure.}

\item{relative_risk_stratum_1}{The relative risk of the outcome
associated with the first stratum.}
}
\value{
A data frame.
}
\description{
Simulate a data set according to Zou (2004, p 703)
}
\examples{
probability_exposed <- 0.5
probability_stratum_1_given_exposed <- 0.6
probability_stratum_1_given_unexposed <- 0.4
risk_given_unexposed_stratum_2 <- 0.2
sim_data <- simulate_data(
  n = 100,
  probability_exposed,
  probability_stratum_1_given_exposed,
  probability_stratum_1_given_unexposed,
  risk_given_unexposed_stratum_2 = 0.2,
  relative_risk_exposed = 2,
  relative_risk_stratum_1 = 1.5
)
}
